+++ src/zensols/datdesc/renderlatex.py
@@ -1,11 +1,20 @@
 """Things rendered for LaTeX files (usually ``.sty``).
 
 """
-from dataclasses import dataclass, field
-from io import StringIO
+from typing import Any, ClassVar
 from collections.abc import Sequence
-from . import DataDescriptionError
+from dataclasses import dataclass, field
+import logging
+import sys
+from itertools import chain
+from datetime import datetime
+import re
+from io import StringIO, TextIOBase
+from zensols.config import Dictable
+from .domain import DataDescriptionError
 from .render import RenderableArtifact
+
+logger = logging.getLogger(__name__)
 
 
 @dataclass
@@ -14,6 +23,12 @@ class RenderableLatexArtifact(RenderableArtifact):
     utility method for generating latex commands.
 
     """
+    _FILE_NAME_REGEX: ClassVar[re.Pattern] = re.compile(r'(.+)\.yml')
+    """Used to narrow down to a :obj:`package_name`."""
+
+    uses: list[str] = field(default_factory=list)
+    """Comma separated list of packages to use."""
+
     default_params: Sequence[Sequence[str]] = field(default_factory=list)
     """Default parameters to be substituted in the template that are
     interpolated by (i.e. LaTeX) numeric values such as #1, #2, etc.  This is a
@@ -27,6 +42,26 @@ class RenderableLatexArtifact(RenderableArtifact):
     :obj:`default_params`.
 
     """
+    def __post_init__(self):
+        super().__post_init__()
+        if isinstance(self.uses, str):
+            self.uses = re.split(r'\s*,\s*', self.uses)
+
+    @property
+    def package_name(self) -> str:
+        """Return the package name for the table in ``table_path``."""
+        fname: str = self.definition_file.name
+        m: re.Match = self._FILE_NAME_REGEX.match(fname)
+        if m is None:
+            raise DataDescriptionError(
+                f'Does not appear to be a YAML file: {fname}', self.name)
+        return m.group(1)
+
+    def _get_template_params(self) -> dict[str, Any]:
+        params: dict[str, Any] = dict(self.asdict())
+        params.update(self._get_command_params())
+        return params
+
     def _get_command_params(self) -> dict[str, str]:
         """Create parameters prefixed as a nested :class:`~builtins.dict` with
         name ``p`` to be substituted as values in the table template.  A
@@ -74,3 +109,49 @@ class RenderableLatexArtifact(RenderableArtifact):
         aparams['argdef'] = proto
         params['usage'] = usage.getvalue()
         return params
+
+
+@dataclass
+class RenderableLatexPackage(Dictable):
+    """Generate a Latex table from a CSV file.
+
+    """
+    artifacts: Sequence[RenderableLatexArtifact] = field()
+    """A list of instances to create Latex definitions."""
+
+    name: str = field()
+    """The name Latex .sty package."""
+
+    description: str = '{date} Artifacts'
+    """The package description."""
+
+    def _write_header(self, depth: int, writer: TextIOBase):
+        date = datetime.now().strftime('%Y/%m/%d')
+        params: dict[str, Any] = self.asdict() | {'date': date}
+        desc: str = self.description.format(**params)
+        writer.write("""\\NeedsTeXFormat{LaTeX2e}
+\\ProvidesPackage{%(package_name)s}[%(desc)s]
+
+""" % {'package_name': self.name, 'desc': desc})
+        uses: set[str] = set(chain.from_iterable(
+            map(lambda t: t.uses, self.artifacts)))
+        for use in sorted(uses):
+            writer.write(f'\\usepackage{{{use}}}\n')
+        if len(uses) > 0:
+            writer.write('\n')
+
+    def write(self, depth: int = 0, writer: TextIOBase = sys.stdout):
+        """Write the Latex table to the writer given in the initializer.
+
+        """
+        art_len: int = len(self.artifacts)
+        self._write_header(depth, writer)
+        for i, table in enumerate(self.artifacts):
+            try:
+                table.write(depth, writer)
+            except Exception as e:
+                msg: str = f"could not format table '{table.name}': {e}"
+                self._write_line(f'% erorr: {msg}', depth, writer)
+                logger.error(msg, e)
+            if i < art_len:
+                writer.write('\n')