+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Value formatting for bar plots.
 - A base class `RenderableArtifact` created by `Renderable`.
+- Figure output adds Latex sty files.
 
 
 ## [1.4.8] - 2026-06-30